runfix: Fix sync between live self user and DB data (#15046)
